requirements.txt
Pillow
customtkinter
CTkMenuBar
